beta 0.3

+++ ncel/models/featureGenerator.py
@@ -150,7 +150,7 @@ class FeatureGenerator:
             for i, c in enumerate(m.candidates):
                 candidate_ids.append(c.id)
                 gold_ids.append([1,0] if c.getIsGlod() else [0,1])
-        return np.array(gold_ids), np.array(candidate_ids)
+        return np.array(gold_ids), np.array(candidate_ids, dtype=np.int32)
 
     def getFeatures(self, doc):
         # doc.n_candidates * base_feature_dim
